fix(utxo-lib): refactor imports to remove circular dependencies
Issue: BTC-1966
